refactor the change to the upper class
